# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: InspectProof API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List reports
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
    post:
      operationId: generateReport
      tags: [reports]
      summary: Generate report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateReportRequest"
      responses:
        "201":
          description: Report generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /reports/{id}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportDetail"
...
components:
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        siteAddress:
          type: string
        suburb:
          type: string
        state:
          type: string
        postcode:
          type: string
        clientName:
          type: string
        builderName:
          type: string
          nullable: true
        designerName:
          type: string
          nullable: true
        daNumber:
          type: string
          nullable: true
        certificationNumber:
          type: string
          nullable: true
        buildingClassification:
          type: string
        projectType:
          type: string
          enum: [residential, commercial, industrial, mixed_use, infrastructure]
        status:
          type: string
          enum: [active, on_hold, completed, archived]
        stage:
          type: string
          enum: [pre_construction, footings, slab, frame, lock_up, fit_out, final, completed]
        assignedCertifierId:
          type: integer
          nullable: true
        assignedInspectorId:
          type: integer
          nullable: true
        startDate:
          type: string
          nullable: true
        expectedCompletionDate:
          type: string
          nullable: true
        completedDate:
          type: string
          nullable: true
        totalInspections:
          type: integer
        openIssues:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - siteAddress
        - suburb
        - state
        - postcode
        - clientName
        - buildingClassification
        - projectType
        - status
        - stage
        - totalInspections
        - openIssues
        - createdAt
        - updatedAt
...
    Inspection:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        projectName:
          type: string
        inspectionType:
          type: string
          enum: [footings, slab, frame, final, fire_safety, pool_barrier, special, preliminary, progress]
        status:
          type: string
          enum: [scheduled, in_progress, completed, follow_up_required, cancelled]
        scheduledDate:
          type: string
        scheduledTime:
          type: string
          nullable: true
        completedDate:
          type: string
          nullable: true
        inspectorId:
          type: integer
          nullable: true
        inspectorName:
          type: string
          nullable: true
        duration:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
        weatherConditions:
          type: string
          nullable: true
        checklistTemplateId:
          type: integer
          nullable: true
        passCount:
          type: integer
        failCount:
          type: integer
        naCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - projectId
        - projectName
        - inspectionType
        - status
        - scheduledDate
        - passCount
        - failCount
        - naCount
        - createdAt
...
    Issue:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        inspectionId:
          type: integer
          nullable: true
        title:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        status:
          type: string
          enum: [open, in_progress, resolved, closed, deferred]
        location:
          type: string
          nullable: true
        codeReference:
          type: string
          nullable: true
        responsibleParty:
          type: string
          nullable: true
        dueDate:
          type: string
          nullable: true
        resolvedDate:
          type: string
          nullable: true
        assignedToId:
          type: integer
          nullable: true
        projectName:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - projectId
        - title
        - description
        - severity
        - status
        - projectName
        - createdAt
        - updatedAt
...
    Report:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        inspectionId:
          type: integer
          nullable: true
        title:
          type: string
        reportType:
          type: string
          enum: [summary, detailed, defect_notice, follow_up]
        status:
          type: string
          enum: [draft, final, sent]
        generatedById:
          type: integer
        generatedByName:
          type: string
        projectName:
          type: string
        createdAt:
          type: string
      required:
        - id
        - projectId
        - title
        - reportType
        - status
        - generatedById
        - generatedByName
        - projectName
        - createdAt
    ReportDetail:
      allOf:
        - $ref: "#/components/schemas/Report"
        - type: object
          properties:
            project:
              $ref: "#/components/schemas/Project"
            inspection:
              $ref: "#/components/schemas/Inspection"
              nullable: true
            issues:
              type: array
              items:
                $ref: "#/components/schemas/Issue"
          required:
            - project
            - issues
    GenerateReportRequest:
      type: object
      properties:
        projectId:
          type: integer
        inspectionId:
          type: integer
          nullable: true
        title:
          type: string
        reportType:
          type: string
          enum: [summary, detailed, defect_notice, follow_up]
        generatedById:
          type: integer
      required:
        - projectId
        - title
        - reportType
        - generatedById
...